test_requirements.txt
django-libs>=1.67.2
flake8
django-nose
coverage
django-coverage
ipdb
fabric3
mixer
mock
tox
